% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsubset.R
\name{LMsubset}
\alias{LMsubset}
\title{Subset Likelihood Modeling for Logistic Regression}
\usage{
LMsubset(formula = formula, data = data, size, burnin, conf_level,
  approx_method = approx_method)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are the same to the formula in lm.}

\item{data}{a data frame containing the variables in the model. If not found in data, the variables are taken from environment(formula).}

\item{size}{the number of MCMC iterations saved (target distribution is the likelihood function of parameters in the logistic regression)}

\item{burnin}{the number of MCMC iterations discarded.}

\item{conf_level}{a vector which consists of levels of credible intervals}

\item{approx_method}{the method to approximate the posterior distribution such as normal ("Norm") or skew normal ("SN"), the default one is normal distribution}
}
\value{
\item{prob_compare}{a dataframe with two columns: approximate probability and true proability. The probability under different credible regions defined by conf_levels is estimated by using Monte Carlo methods}

\item{norm.mean}{mean parameter of fitted normal distribution}

\item{norm.var}{variance (covariance) of fitted normal distribution}

\item{sn.xi}{location parameter of fitted skew normal distribution if approx_method is "SN", Null otherwise}

\item{sn.omega}{scale parameter of fitted skew normal distributionif approx_method is "SN", Null otherwise}

\item{sn.alpha}{shape parameter of fitted skew normal distributionif approx_method is "SN", Null otherwise}
}
\description{
Model the posterior distribution of parameters (likelihood function) in logistic regression using the normal distribution or skew normal distribution, where the prior distribution is the uniform distribution.
}
\details{
Fit logistic regression to data with formula, simulate a sample of size = size with burnin = burnin using Monte carol methods from the posterior  distribution (likelihood function) of coefficients. There are two approximate methods considered in this function: Normal approximation and Skew-normal approximation.  
}
\examples{
x <- matrix(rnorm(1000*5), ncol=5)
y <- rbinom(1000,1,0.5)
df <- as.data.frame(x)
names(df) <- paste("x",1:5,sep="")
df$y <- y
rst <- LMsubset(y~x1+x2, data =df, size=500, burnin =50, conf_level = seq(0.05, 0.95, 0.05), approx_method = "SN")
}
\seealso{
\code{\link{subset_approx}}
}
\author{
Qi Liu
}
