% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predNew.local.R
\name{predNew.local}
\alias{predNew.local}
\title{Fitted Values at New Data (in Memory) Based on the Fitted Results from Likelihood Modeling}
\usage{
predNew.local(fitted_par, formula, data, size = 1000)
}
\arguments{
\item{fitted_par}{object returned by drml function}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are the same to the formula in lm.}
\item{data}{a data frame containing the variables in the model. If not found in data, the variables are taken from environment(formula).}
\item{size}{the number of samples drawn from the distribution of model parameters}
}
\value{
0.025, 0.5, 0.975 quantiles of the distribution of fitted predict probability
}
\description{
predNew.local is a function to provide the 0.025, 0.5, 0.975 quantiles of the distribution of fitted predict probability based on the fitted density of model parameters.
}
\examples{
set.seed(100)
library(datadr)
x <- matrix(rnorm(1000*5), ncol=5)
ttheta <- rep(1,5)
y <- rbinom(1000, 1, 1 / (1 + exp(- x \%*\% ttheta)))
df <- cbind(y,x)
df <- as.data.frame(df)
names(df) <- c("y", "x1","x2","x3","x4","x5")
df_ddf <- ddf(df)
df_div <- divide(df_ddf, by =rrDiv(500))
rst<- drml(df_div, y~x1+x2, size =1000, burnin =50, approx_method = "SN") 
pred <- predNew.local(rst, y~x1+x2, df, 1000)
}
