% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_approx.R
\name{subset_approx}
\alias{subset_approx}
\title{Likelihood Modeling for Logistic Regression in subset sense}
\usage{
subset_approx(x, y, size, burnin, conf_level, approx_method = approx_method)
}
\arguments{
\item{x}{the model matrix}

\item{y}{the response variable}

\item{size}{the number of MCMC iterations saved (target distribution is the posterior distribution of parameters in the logistic regression)}

\item{burnin}{the number of MCMC iterations discarded.}

\item{conf_level}{a vector which consists of levels of credible intervals}

\item{approx_method}{the method to approximate the posterior distribution such as normal or skew normal, the default one is normal distribution}
}
\value{
\item{prob_compare}{a dataframe with two columns: approximate probability and true proability. The probability under different credible regions defined by conf_levels is estimated by using Monte Carlo methods}

\item{norm.mean}{mean parameter of fitted normal distribution}

\item{norm.var}{variance (covariance) of fitted normal distribution}

\item{sn.xi}{location parameter of fitted skew normal distribution if approx_method is "SN", Null otherwise}

\item{sn.omega}{scale parameter of fitted skew normal distributionif approx_method is "SN", Null otherwise}

\item{sn.alpha}{shape parameter of fitted skew normal distributionif approx_method is "SN", Null otherwise}
}
\description{
Model the posterior distribution of parameters (likelihood function) in logistic regression using the normal distribution or skew normal distribution, where the prior distribution is the uniform distribution.
}
\examples{

x <- matrix(rnorm(1000*5), ncol=5)
y <- rbinom(1000,1,0.5)
a <- subset_approx(x,y, size=5000, burnin =500, conf_level = seq(0.05, 0.95, 0.05), approx_method = "Norm")
}
\author{
Qi Liu
}
